+++ src/pipes/index.ts
@@ -1,4 +1,5 @@
 export * from './addNewOption/addNewOption.pipe';
+export * from './displayOptionValue/displayOptionValue.pipe';
 export * from './displayValue/displayValue.pipe';
 export * from './getPlugin/getPlugin.pipe';
 export * from './groupedListOptions/groupedListOptions.pipe';